Make the build script a bit more robust and friendly in face of errors (#20)
* Check if build_bin dir and ld symlink exist before trying to make them

If a previous build failed then the directory and symlink may already exist in
"$BUILD_DIR". If we try to make them again then we fail _differently_ on
subsequent build attempts, which is confusing.

* Check for clang in path before attempting to do build

If there's no clang in the path then the later call to configure exits
confusingly with an error like "error: C compiler cannot create
executables". That's technically correct, but not very helpful.

Co-authored-by: Eric Ridge <eebbrr@gmail.com>